refactor(core): split frame-stats and scale-axis modules
Dispatch non-identity LayerFrame stats into binning (count/bin/density) and
fit (smooth/boxplot/summary) modules; split positional scale train vs evidence
collection. Characterization tests pin density/smooth/boxplot/summary frames
and axis evidence for bars and bins.
